NF: Create SyncHandler
This class deals with the entire Sync part of the deck picker.
