+++ AnkiDroid/src/main/java/com/ichi2/anki/MyAccount.kt
@@ -35,15 +35,18 @@ import androidx.core.content.ContextCompat
 import androidx.core.view.isVisible
 import com.google.android.material.textfield.TextInputEditText
 import com.google.android.material.textfield.TextInputLayout
+import com.ichi2.anki.CollectionManager.withCol
 import com.ichi2.anki.dialogs.help.HelpDialog
 import com.ichi2.anki.pages.RemoveAccountFragment
 import com.ichi2.anki.preferences.sharedPrefs
 import com.ichi2.anki.utils.ext.removeFragmentFromContainer
 import com.ichi2.anki.utils.ext.showDialogFragment
+import com.ichi2.libanki.syncLogin
 import com.ichi2.ui.TextInputEditField
 import com.ichi2.utils.AdaptionUtil.isUserATestClient
 import com.ichi2.utils.KotlinCleanup
 import com.ichi2.utils.Permissions
+import net.ankiweb.rsdroid.exceptions.BackendSyncException
 import timber.log.Timber
 
 /**
@@ -321,6 +324,37 @@ open class MyAccount : AnkiActivity() {
     private fun openAnkiDroidPrivacyPolicy() {
         Timber.i("Opening 'Privacy policy'")
         showDialogFragment(HelpDialog.newPrivacyPolicyInstance())
+    }
+
+    fun handleNewLogin(
+        username: String,
+        password: String,
+        resultLauncher: ActivityResultLauncher<String>,
+    ) {
+        val endpoint = getEndpoint(this)
+        launchCatchingTask {
+            val auth =
+                try {
+                    withProgress(
+                        extractProgress = {
+                            text = getString(R.string.sign_in)
+                        },
+                        onCancel = ::cancelSync,
+                    ) {
+                        withCol {
+                            syncLogin(username, password, endpoint)
+                        }
+                    }
+                } catch (exc: BackendSyncException.BackendSyncAuthFailedException) {
+                    // auth failed; clear out login details
+                    updateLogin(baseContext, "", "")
+                    throw exc
+                }
+            updateLogin(baseContext, username, auth.hkey)
+            setResult(RESULT_OK)
+            checkNotificationPermission(this@MyAccount, resultLauncher)
+            finish()
+        }
     }
 
     companion object {
