+++ AnkiDroid/src/main/java/com/ichi2/anki/dialogs/BackupPromptDialog.kt
@@ -164,7 +164,7 @@ class BackupPromptDialog private constructor(
             backupPrompt.apply {
                 build(isLoggedIn) {
                     if (isLoggedIn) {
-                        deckPicker.sync(conflict = null)
+                        deckPicker.syncHandler.sync(conflict = null)
                     } else {
                         deckPicker.exportCollection()
                     }
